+++ subsys/net/lib/lwm2m/lwm2m_message_handling.h
@@ -73,4 +73,6 @@ int lwm2m_write_handler(struct lwm2m_engine_obj_inst *obj_inst, struct lwm2m_eng
 enum coap_block_size lwm2m_default_block_size(void);
 
 int lwm2m_parse_peerinfo(char *url, struct lwm2m_ctx *client_ctx, bool is_firmware_uri);
+void lwm2m_clear_block_contexts(void);
+
 #endif /* LWM2M_MESSAGE_HANDLING_H */
