+++ subsys/net/lib/lwm2m/lwm2m_util.h
@@ -27,4 +27,22 @@ uint16_t lwm2m_atou16(const uint8_t *buf, uint16_t buflen, uint16_t *len);
 int lwm2m_string_to_path(const char *pathstr, struct lwm2m_obj_path *path, char delim);
 
 bool lwm2m_obj_path_equal(const struct lwm2m_obj_path *a, const struct lwm2m_obj_path *b);
+
+/**
+ * @brief Used for debugging to print ip addresses.
+ *
+ * @param addr sockaddr for socket using ipv4 or ipv6
+ * @return ip address in readable form
+ */
+char *lwm2m_sprint_ip_addr(const struct sockaddr *addr);
+
+/**
+ * @brief Converts the token to a printable format.
+ *
+ * @param[in] token Token to be printed
+ * @param[in] tkl Lenghts of token
+ * @return char buffer with the string representation of the token
+ */
+char *sprint_token(const uint8_t *token, uint8_t tkl);
+
 #endif /* LWM2M_UTIL_H_ */
