+++ tests/net/lib/lwm2m/lwm2m_engine/src/stubs.h
@@ -53,6 +53,7 @@ DECLARE_FAKE_VALUE_FUNC(struct lwm2m_engine_obj_field *, lwm2m_get_engine_obj_fi
 DECLARE_FAKE_VALUE_FUNC(int, lwm2m_get_bool, const struct lwm2m_obj_path *, bool *);
 DECLARE_FAKE_VALUE_FUNC(sys_slist_t *, lwm2m_engine_obj_inst_list);
 DECLARE_FAKE_VALUE_FUNC(int, lwm2m_delete_obj_inst, uint16_t, uint16_t);
+DECLARE_FAKE_VOID_FUNC(lwm2m_clear_block_contexts);
 
 #define DO_FOREACH_FAKE(FUNC)                                                                      \
 	do {                                                                                       \
@@ -79,6 +80,7 @@ DECLARE_FAKE_VALUE_FUNC(int, lwm2m_delete_obj_inst, uint16_t, uint16_t);
 		FUNC(lwm2m_get_bool)                                                               \
 		FUNC(lwm2m_engine_obj_inst_list)                                                   \
 		FUNC(lwm2m_delete_obj_inst)                                                        \
+		FUNC(lwm2m_clear_block_contexts)                                                   \
 	} while (0)
 
 #endif /* STUBS_H */
